chore: move Relativize trait to separate relativize.rs file
